# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * India Cafe API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Location {
  id: number;
  name: string;
  address: string;
  phone: string;
  hours: string;
  latitude?: number | null;
  longitude?: number | null;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  sortOrder: number;
}

export interface CreateCategoryBody {
  name: string;
  slug: string;
  sortOrder?: number;
}

export interface MenuItem {
  id: number;
  name: string;
  description: string;
  /** Decimal string e.g. '14.99' */
  price: string;
  imageUrl?: string | null;
  categoryId: number;
  categoryName?: string | null;
  isVegetarian: boolean;
  isVegan: boolean;
  isGlutenFree: boolean;
  /** 0-3 (mild..hot) */
  spiceLevel: number;
  isAvailable: boolean;
  isFeatured: boolean;
}

export interface CreateMenuItemBody {
  name: string;
  description: string;
  price: string;
  imageUrl?: string | null;
  categoryId: number;
  isVegetarian?: boolean;
  isVegan?: boolean;
  isGlutenFree?: boolean;
  spiceLevel?: number;
  isAvailable?: boolean;
  isFeatured?: boolean;
}

export interface UpdateMenuItemBody {
  name?: string;
  description?: string;
  price?: string;
  imageUrl?: string | null;
  categoryId?: number;
  isVegetarian?: boolean;
  isVegan?: boolean;
  isGlutenFree?: boolean;
  spiceLevel?: number;
  isAvailable?: boolean;
  isFeatured?: boolean;
}

export interface Testimonial {
  id: number;
  name: string;
  rating: number;
  message: string;
  isApproved: boolean;
  createdAt: string;
}

export interface CreateTestimonialBody {
  name: string;
  rating: number;
  message: string;
}

export interface UpdateTestimonialBody {
  isApproved?: boolean;
}

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  phone?: string | null;
  message: string;
  createdAt: string;
}

export interface CreateContactBody {
  name: string;
  email: string;
  phone?: string | null;
  message: string;
}

export interface OrderItem {
  id: number;
  menuItemId: number;
  name: string;
  price: string;
  quantity: number;
  notes?: string | null;
}

export interface Order {
  id: number;
  orderNumber: string;
  /** pending_payment | paid | preparing | ready | completed | cancelled */
  status: string;
  /** pickup | delivery */
  orderType: string;
  locationId: number;
  locationName?: string | null;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  deliveryAddress?: string | null;
  scheduledFor?: string | null;
  notes?: string | null;
  subtotal: string;
  tax: string;
  tipAmount: string;
  deliveryFee: string;
  total: string;
  paymentMethod?: string | null;
  paypalOrderId?: string | null;
  items: OrderItem[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateOrderItem {
  menuItemId: number;
  quantity: number;
  notes?: string | null;
}

export interface CreateOrderBody {
  /** pickup | delivery */
  orderType: string;
  locationId: number;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  deliveryAddress?: string | null;
  scheduledFor?: string | null;
  notes?: string | null;
  tipAmount?: string | null;
  items: CreateOrderItem[];
}

export interface UpdateOrderBody {
  status?: string;
}

export interface PaypalCreateOrderBody {
  orderId: number;
}

export interface PaypalOrderResponse {
  id: string;
}

export interface PaypalCaptureResponse {
  success: boolean;
  orderId: number;
  orderNumber: string;
}

export interface PaypalSetupResponse {
  clientId: string;
}

export interface AdminLoginBody {
  password: string;
}

export interface AdminLoginResponse {
  token: string;
}

export interface AdminMeResponse {
  authenticated: boolean;
}

export type DashboardSummaryStatusBreakdownItem = {
  status: string;
  count: number;
};

export type DashboardSummaryTopItemsItem = {
  menuItemId: number;
  name: string;
  quantity: number;
  revenue: string;
};

export interface DashboardSummary {
  ordersToday: number;
  revenueToday: string;
  ordersThisWeek: number;
  revenueThisWeek: string;
  pendingOrders: number;
  preparingOrders: number;
  readyOrders: number;
  totalMenuItems: number;
  totalCustomers: number;
  statusBreakdown: DashboardSummaryStatusBreakdownItem[];
  recentOrders: Order[];
  topItems: DashboardSummaryTopItemsItem[];
}

export type ListMenuItemsParams = {
  categoryId?: number;
  search?: string;
  featured?: boolean;
};

export type LookupOrderParams = {
  orderNumber: string;
  email: string;
};

export type AdminListOrdersParams = {
  status?: string;
  orderType?: string;
};