# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: India Cafe API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /menu/items:
    get:
      operationId: listMenuItems
      tags: [menu]
      summary: List menu items
      parameters:
        - in: query
          name: categoryId
          schema: { type: integer }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: featured
          schema: { type: boolean }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
  /menu/items/{id}:
    get:
      operationId: getMenuItem
      tags: [menu]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
...
  /admin/orders:
    get:
      operationId: adminListOrders
      tags: [admin]
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: orderType
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /admin/orders/{id}:
    get:
      operationId: adminGetOrder
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: adminUpdateOrder
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /admin/menu/items:
    get:
      operationId: adminListMenuItems
      tags: [admin]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: adminCreateMenuItem
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMenuItemBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
  /admin/menu/items/{id}:
    patch:
      operationId: adminUpdateMenuItem
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMenuItemBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
    delete:
      operationId: adminDeleteMenuItem
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /admin/menu/categories:
    post:
      operationId: adminCreateCategory
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /admin/menu/categories/{id}:
    delete:
      operationId: adminDeleteCategory
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
  /admin/testimonials/{id}:
    patch:
      operationId: adminUpdateTestimonial
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTestimonialBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"
    delete:
      operationId: adminDeleteTestimonial
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    Category:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        sortOrder: { type: integer }
      required: [id, name, slug, sortOrder]
    CreateCategoryBody:
      type: object
      properties:
        name: { type: string }
        slug: { type: string }
        sortOrder: { type: integer }
      required: [name, slug]
    MenuItem:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string }
        price: { type: string, description: "Decimal string e.g. '14.99'" }
        imageUrl: { type: string, nullable: true }
        categoryId: { type: integer }
        categoryName: { type: string, nullable: true }
        isVegetarian: { type: boolean }
        isVegan: { type: boolean }
        isGlutenFree: { type: boolean }
        spiceLevel: { type: integer, description: "0-3 (mild..hot)" }
        isAvailable: { type: boolean }
        isFeatured: { type: boolean }
      required: [id, name, description, price, categoryId, isVegetarian, isVegan, isGlutenFree, spiceLevel, isAvailable, isFeatured]
    CreateMenuItemBody:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        price: { type: string }
        imageUrl: { type: string, nullable: true }
        categoryId: { type: integer }
        isVegetarian: { type: boolean }
        isVegan: { type: boolean }
        isGlutenFree: { type: boolean }
        spiceLevel: { type: integer }
        isAvailable: { type: boolean }
        isFeatured: { type: boolean }
      required: [name, description, price, categoryId]
    UpdateMenuItemBody:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        price: { type: string }
        imageUrl: { type: string, nullable: true }
        categoryId: { type: integer }
        isVegetarian: { type: boolean }
        isVegan: { type: boolean }
        isGlutenFree: { type: boolean }
        spiceLevel: { type: integer }
        isAvailable: { type: boolean }
        isFeatured: { type: boolean }
    Testimonial:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        rating: { type: integer }
        message: { type: string }
        isApproved: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, name, rating, message, isApproved, createdAt]
...
    UpdateTestimonialBody:
      type: object
      properties:
        isApproved: { type: boolean }
...
    OrderItem:
      type: object
      properties:
        id: { type: integer }
        menuItemId: { type: integer }
        name: { type: string }
        price: { type: string }
        quantity: { type: integer }
        notes: { type: string, nullable: true }
      required: [id, menuItemId, name, price, quantity]
    Order:
      type: object
      properties:
        id: { type: integer }
        orderNumber: { type: string }
        status: { type: string, description: "pending_payment | paid | preparing | ready | completed | cancelled" }
        orderType: { type: string, description: "pickup | delivery" }
        locationId: { type: integer }
        locationName: { type: string, nullable: true }
        customerName: { type: string }
        customerEmail: { type: string }
        customerPhone: { type: string }
        deliveryAddress: { type: string, nullable: true }
        scheduledFor: { type: string, format: date-time, nullable: true }
        notes: { type: string, nullable: true }
        subtotal: { type: string }
        tax: { type: string }
        tipAmount: { type: string }
        deliveryFee: { type: string }
        total: { type: string }
        paymentMethod: { type: string, nullable: true }
        paypalOrderId: { type: string, nullable: true }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, orderNumber, status, orderType, locationId, customerName, customerEmail, customerPhone, subtotal, tax, tipAmount, deliveryFee, total, items, createdAt, updatedAt]
...
    UpdateOrderBody:
      type: object
      properties:
        status: { type: string }
...
    AdminLoginBody:
      type: object
      properties:
        password: { type: string }
      required: [password]
    AdminLoginResponse:
      type: object
      properties:
        token: { type: string }
      required: [token]
...